+++ analysis/import/SCMLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/scmlogparser/SCMLogProjectCreator.kt
@@ -12,15 +12,15 @@ import java.util.stream.Stream
  */
 class SCMLogProjectCreator(
         parserStrategy: LogParserStrategy,
-        metricsFactory: MetricsFactory,
+        private val metricsFactory: MetricsFactory,
         private val projectConverter: ProjectConverter,
-        private val silent: Boolean = false
+        silent: Boolean = false
 ) {
 
     private val logLineParser: LogLineParser = LogLineParser(parserStrategy, metricsFactory, silent)
 
     fun parse(lines: Stream<String>): Project {
         val versionControlledFiles = logLineParser.parse(lines)
-        return projectConverter.convert(versionControlledFiles)
+        return projectConverter.convert(versionControlledFiles, metricsFactory)
     }
 }