+++ analysis/import/SCMLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/scmlogparser/input/metrics/Metric.kt
@@ -2,6 +2,7 @@ package de.maibornwolff.codecharta.importer.scmlogparser.input.metrics
 
 import de.maibornwolff.codecharta.importer.scmlogparser.input.Commit
 import de.maibornwolff.codecharta.importer.scmlogparser.input.Modification
+import de.maibornwolff.codecharta.model.AttributeType
 import de.maibornwolff.codecharta.model.Edge
 
 interface Metric {
@@ -19,6 +20,15 @@ interface Metric {
     fun getEdges(): List<Edge> {
         return listOf()
     }
+
+    fun attributeType(): AttributeType {
+        return AttributeType.absolute
+    }
+
+    fun edgeAttributeType(): AttributeType? {
+        return null
+    }
+
 
     fun registerModification(modification: Modification) {
         // defaults to: do nothing