+++ visualization/app/codeCharta/ui/codeMap/codeMap.preRender.service.ts
@@ -19,6 +19,7 @@ import { SearchPanelModeActions } from "../../state/store/appSettings/searchPane
 import { isActionOfType } from "../../util/reduxHelper"
 import { SortingOrderAscendingActions } from "../../state/store/appSettings/sortingOrderAscending/sortingOrderAscending.actions"
 import { SortingOptionActions } from "../../state/store/dynamicSettings/sortingOption/sortingOption.actions"
+import { IsAttributeSideBarVisibleActions } from "../../state/store/appSettings/isAttributeSideBarVisible/isAttributeSideBarVisible.actions"
 
 const clone = require("rfdc")()
 
@@ -66,7 +67,8 @@ export class CodeMapPreRenderService implements StoreSubscriber, MetricServiceSu
 			!isActionOfType(actionType, IsLoadingFileActions) &&
 			!isActionOfType(actionType, SearchPanelModeActions) &&
 			!isActionOfType(actionType, SortingOrderAscendingActions) &&
-			!isActionOfType(actionType, SortingOptionActions)
+			!isActionOfType(actionType, SortingOptionActions) &&
+			!isActionOfType(actionType, IsAttributeSideBarVisibleActions)
 		) {
 			this.debounceRendering()
 		}
@@ -95,20 +97,17 @@ export class CodeMapPreRenderService implements StoreSubscriber, MetricServiceSu
 	}
 
 	private decorateIfPossible() {
-		if (
-			this.unifiedMap &&
-			this.storeService.getState().files.fileStatesAvailable() &&
-			this.unifiedFileMeta &&
-			this.metricService.getMetricData()
-		) {
+		const state = this.storeService.getState()
+		if (this.unifiedMap && state.files.fileStatesAvailable() && this.unifiedFileMeta && this.metricService.getMetricData()) {
 			NodeDecorator.decorateMap(this.unifiedMap, this.unifiedFileMeta, this.metricService.getMetricData())
 			this.getEdgeMetricsForLeaves(this.unifiedMap)
-			NodeDecorator.decorateParentNodesWithSumAttributes(
+			NodeDecorator.decorateParentNodesWithAggregatedAttributes(
 				this.unifiedMap,
-				this.storeService.getState().fileSettings.blacklist,
+				state.fileSettings.blacklist,
 				this.metricService.getMetricData(),
 				this.edgeMetricDataService.getMetricData(),
-				this.storeService.getState().files.isDeltaState()
+				state.files.isDeltaState(),
+				state.fileSettings.attributeTypes
 			)
 		}
 	}